Authentication

+++ api/rendez-vous/delete.php
@@ -8,27 +8,33 @@
     
     include_once '../../config/Database.php';
     include_once '../../models/RendezVous.php';
+    include_once '../../Token/token.php';
 
-    // Instantiate DB & connect 
-    $database = new Database(); 
-    $db = $database->connect(); 
-
-    // Instantiate rdv object 
-    $rdv = new RendezVous($db); 
-
-    // Get raw rdv data 
-    $data = json_decode(file_get_contents("php://input"));
-
-    // Set ID to delete
-    $rdv->id = $data->id;
-
-    // DELETE rdv
-    if($rdv->delete()){
-        echo json_encode(
-            array('message'=>'RDV Deleted')
-        );
-    }   else {
-        echo json_encode(
-            array('message'=>'RDV Not Deleted')
-        );
+    if (Token::verifier()) {
+        // Instantiate DB & connect 
+        $database = new Database(); 
+        $db = $database->connect(); 
+    
+        // Instantiate rdv object 
+        $rdv = new RendezVous($db); 
+    
+        // Get raw rdv data 
+        $data = json_decode(file_get_contents("php://input"));
+    
+        // Set ID to delete
+        $rdv->id = $data->id;
+    
+        // DELETE rdv
+        if($rdv->delete()){
+            echo json_encode(
+                array('message'=>'RDV Deleted')
+            );
+        }   else {
+            echo json_encode(
+                array('message'=>'RDV Not Deleted')
+            );
+        }
+    } else {
+        echo json_encode(array('message' => 'invalide token'));
     }
+
